fix: address PR review — write order, thread safety, code quality
- Fix #1: Set LastStateChangeAt on initial registration
- Fix #2: Remove meaningless Pending→Pending audit entry at registration
- Fix #3: MongoDB first, Redis second (correct write-through order)
- Fix #4: Thread-safe index creation with SemaphoreSlim + volatile
- Fix #5: WriteAuditEntryAsync visibility reduced to private
- Fix #6: ValidatorDocument extracted to Models/ValidatorDocument.cs
- Fix #7: Enum.TryParse with Suspended fallback for unknown status values
- Fix #8: CancellationToken added to PersistToMongoAsync
- Fix #11: WalletSequence mutable fields documented
